fixed bugs with submitting remove btn w/o checking anything, bug with remove method

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -2,8 +2,10 @@ package org.launchcode.controllers;
 
 import org.launchcode.models.Category;
 import org.launchcode.models.Cheese;
+import org.launchcode.models.Menu;
 import org.launchcode.models.data.CategoryDao;
 import org.launchcode.models.data.CheeseDao;
+import org.launchcode.models.data.MenuDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,6 +13,7 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.List;
 
 /**
  * Created by LaunchCode
@@ -24,6 +27,9 @@ public class CheeseController {
 
     @Autowired
     private CategoryDao categoryDao;
+
+    @Autowired
+    private MenuDao menuDao;
 
     @RequestMapping(value = "")
     public String index(Model model) {
@@ -63,6 +69,10 @@ public class CheeseController {
     @RequestMapping(value = "remove", method = RequestMethod.GET)
     public String displayRemoveCheeseForm(Model model) {
 
+        // need checkedId as a checkbox on our form that is always checked,
+        // in case user submits form without checking a box
+        int checkedId = 0;
+        model.addAttribute("checkedId", checkedId);
         model.addAttribute("cheeses", cheeseDao.findAll());
         model.addAttribute("title", "Remove Cheese");
 
@@ -71,9 +81,31 @@ public class CheeseController {
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
     public String processRemoveCheeseForm(@RequestParam int[] cheeseIds) {
-
+        // iterate over checked cheeses from the form
         for (int cheeseId : cheeseIds) {
-            cheeseDao.delete(cheeseId);
+            if (cheeseId != 0) { // 0 is the default; see displayRemoveCheeseForm
+
+                // some issue w/ bulk delete on the owning side
+                // of a many-to-many relation ---> means:
+                // must delete cheese from menus before completely deleting cheese
+
+                // iterate through all menus
+                for (Menu menu : menuDao.findAll()) {
+
+                    // iterate through cheeses of current menu
+                    List<Cheese> someCheeses = menu.getCheeses();
+                    for (int i = 0; i < someCheeses.size(); i++) {
+
+                        // check current cheese's id against @RequestParam cheeseId
+                        if (someCheeses.get(i).getId() == cheeseId) {
+                            menu.removeItem(someCheeses.get(i));
+                        }
+                    }
+                }
+
+                // only now may you delete the cheese
+                cheeseDao.delete(cheeseId);
+            }
         }
 
         return "redirect:";
